// SDKs/twilio.com/twilio_sync_v1/1.36.0/typescript/src/sdk/models/shared/syncv1servicesyncmapsyncmapitem.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class SyncV1ServiceSyncMapSyncMapItem extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=created_by" })
  createdBy?: string;

  @Metadata({ data: "json, name=data" })
  data?: any;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_expires" })
  dateExpires?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=key" })
  key?: string;

  @Metadata({ data: "json, name=map_sid" })
  mapSid?: string;

  @Metadata({ data: "json, name=revision" })
  revision?: string;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}